+++ libdoa_backup/libdoa/doa.h
@@ -1,5 +1,4 @@
 #include <Eigen/Dense>
-
 #include <math.h>
 #include <stdlib.h>
 
@@ -15,15 +14,15 @@ class aoa_estimator {
 
     // Algorithm constants
     int antenna_num = 16;
-    float distance = 0.04;  // Distância em m
-    float product;
-    float angle_step = 1 * M_PI / 180; // 1º in radians
+    double distance = 0.04;  // Distância em m
+    double product;
+    double angle_step = 1 * M_PI / 180;  // 1º in radians
 
     // Last estimated direction
-    float az_index = 0;
-    float el_index = 0;
-    float az_esprit;  // Azimuth from ESPRIT algorithm
-    float el_esprit;  // Elevation from ESPRIT algorithm
+    double az_index = 0;
+    double el_index = 0;
+    double az_esprit;  // Azimuth from ESPRIT algorithm
+    double el_esprit;  // Elevation from ESPRIT algorithm
 
     // Antenna array output. It is the algorithm input
     Eigen::Matrix<std::complex<double>, 16, 4> x;
@@ -33,10 +32,10 @@ class aoa_estimator {
     Eigen::Matrix<std::complex<double>, 16, 15> Qn;  // Noise eigenvectors. MUSIC related
     Eigen::Matrix<std::complex<double>, 16, 1> Qs;   // Signal eigenvector
 
-    float x_st;
-    float y_st;
-    std::complex<float> phi_x;
-    std::complex<float> phi_y;
+    double x_st;
+    double y_st;
+    std::complex<double> phi_x;
+    std::complex<double> phi_y;
     Eigen::Matrix<std::complex<double>, 1, 4> a_x;       // steering vector x component
     Eigen::Matrix<std::complex<double>, 1, 4> a_y;       // steering vector y component
     Eigen::Matrix<std::complex<double>, 16, 1> st_vec;   // steering vector
@@ -60,27 +59,27 @@ class aoa_estimator {
 
   public:
     // Init DoA estimator
-    void initDoAEstimator(float elements_distance, int debug_flag);
+    void initDoAEstimator(double elements_distance, int debug_flag);
     // Estimate phase rotation based on CTE reference period IQ data
-    float estimate_phase_rotation(float* i_samples, float* q_samples, int reference_period_length);
-    // Load antenna array output matrix. Converte de float para complexo.
-    void load_x(float** i_samples, float** q_samples, int num_antennas, int num_samples, int debug_flag);
+    double estimate_phase_rotation(double* i_samples, double* q_samples, int reference_period_length);
+    // Load antenna array output matrix. Converte de double para complexo.
+    void load_x(double** i_samples, double** q_samples, int num_antennas, int num_samples, int debug_flag);
     // Compensate the phase rotation due to switching/CTE frequncy offset
-    void compensateRotation(float phase_rotation, int debug_flag);
+    void compensateRotation(double phase_rotation, int debug_flag);
     // Estimate and decomposes covariance matrix
     void estimateRxx(int debug_flag);
     // Update the steering vector
-    void updateSteeringVector(float azimuth, float elevation, int debug_flag);
+    void updateSteeringVector(double azimuth, double elevation, int debug_flag);
     // MUSIC spectrum function. It computes the MUSIC Spatial Spectrum power in the given direction using the Steering Vecotor and Qn matrix
-    float MUSICSpectrum(float azimuth, float elevation, int debug_flag);
+    double MUSICSpectrum(double azimuth, double elevation, int debug_flag);
     // Process MUSIC Spectrum
     void processMUSIC(int azimuth_max, int elevation_max, int debug_flag);
     // Initialize Selecion Matrices for ESPRIT algorithm
     void initSelectionMatrices(int debug_flag);
     // Process ESPRIT algorithm
-    void processESPRIT(float channel_frequency, int debug_flag);
+    void processESPRIT(double channel_frequency, int debug_flag);
     // Get processed values
-    void getProcessed(int algorithm, float* x, float* y);
+    void getProcessed(int algorithm, double* x, double* y);
     // Calculate wavelength based on channel central frequency
-    float calculate_wavelength(float channel_frequency);
+    double calculate_wavelength(double channel_frequency);
 };
